requirements.txt
flask==2.3.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.0
selenium==4.15.0
spacy==3.7.2
transformers==4.35.0
torch==2.1.0
pandas==2.1.0
numpy==1.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.0
sqlite3
smtplib
aiohttp==3.9.0
pydantic==2.5.0
pytest==7.4.0
